030_consumers/python-consumer/requirements.txt
requests
pandas
pyyaml
typing
kafka-python